Separate the log config of background daemons
