+++ sky/utils/schemas.py
@@ -1128,6 +1128,7 @@ _CONTEXT_CONFIG_SCHEMA_KUBERNETES = {
 
 def get_config_schema():
     # pylint: disable=import-outside-toplevel
+    from sky.server import daemons
 
     resources_schema = {
         k: v
@@ -1474,6 +1475,27 @@ def get_config_schema():
         }
     }
 
+    daemon_config = {
+        'type': 'object',
+        'required': [],
+        'additionalProperties': {
+            'type': 'object',
+            'required': [],
+            'additionalProperties': False,
+            'properties': {
+                'log_level': {
+                    'type': 'string',
+                    'case_insensitive_enum': ['DEBUG', 'INFO', 'WARNING'],
+                },
+            }
+        }
+    }
+
+    daemon_schema = {}
+
+    for daemon in daemons.INTERNAL_REQUEST_DAEMONS:
+        daemon_schema[daemon.id] = daemon_config
+
     api_server = {
         'type': 'object',
         'required': [],
@@ -1719,6 +1741,7 @@ def get_config_schema():
             'provision': provision_configs,
             'rbac': rbac_schema,
             'logs': logs_schema,
+            'daemons': daemon_schema,
             **cloud_configs,
         },
     }
